requirements.txt
autopep8==2.0.1
boto3==1.26.61
botocore==1.29.61
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
pycodestyle==2.10.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.11.0
